Fix mistakes, move decommissioning_demand
